tests: update tests for es6 (with profanity)
