+++ js/test_web.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="utf-8"/>
+    <meta charset="utf-8" />
     <title>SoundSwallower tests</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="https://unpkg.com/mocha/mocha.css" />
@@ -10,12 +10,21 @@
     <div id="mocha"></div>
     <script src="https://unpkg.com/chai/chai.js"></script>
     <script src="https://unpkg.com/mocha/mocha.js"></script>
-    <script class="mocha-init">
-      mocha.setup('bdd');
+    <script type="module">
+      mocha.setup("bdd");
     </script>
-    <script src="soundswallower.js"></script>
-    <script src="soundswallower.spec.js"></script>
-    <script class="mocha-exec">
+    <script type="module">
+      import {
+        default as createModule,
+        load_binary_file,
+      } from "./soundswallower.web.js";
+      import { make_tests } from "./tests.js";
+      /* Passing the necessary functions as arguments seems to be the only way
+         to reuse the same test code between Node and the Web, because
+         JavaScript is total fucking shit system. */
+      make_tests(createModule, load_binary_file, chai.assert);
+    </script>
+    <script type="module" class="mocha-exec">
       mocha.run();
     </script>
   </body>
