+++ js/tsconfig.json
@@ -1,10 +1,8 @@
 {
-    "files": [
-        "test_typescript.ts"
-    ],
-    "compilerOptions": {
-        "downlevelIteration": true,
-        "esModuleInterop": true,
-        "lib": ["es2015", "dom"]
-    }
+  "files": ["test_typescript.ts"],
+  "compilerOptions": {
+    "target": "es6",
+    "moduleResolution": "node",
+    "lib": ["es2015", "dom"]
+  }
 }
